+++ src/collective/contact/contactlist/__init__.py
@@ -1,5 +1,7 @@
 # -*- coding: utf-8 -*-
 """Init and utils."""
+import logging
+log = logging.getLogger('collective.contact.contactlist')
 
 from zope.i18nmessageid import MessageFactory